학점(credit) int->float로 변경

+++ app/src/main/java/com/pnuppp/pplusplus/SubjectInfo.java
@@ -2,7 +2,7 @@ package com.pnuppp.pplusplus;
 
 public class SubjectInfo {
     String subjectName;
-    int credit;
+    float credit;
     float grade;
     boolean isMajor;
